:art: make common component
